+++ net_util/src/lib.rs
@@ -32,7 +32,10 @@ use std::{io, mem, net};
 
 pub use mac::{MacAddr, MAC_ADDR_LEN};
 pub use open_tap::{open_tap, Error as OpenTapError};
-pub use queue_pair::{RxVirtio, TxVirtio};
+pub use queue_pair::{
+    NetCounters, NetQueuePair, NetQueuePairError, RxVirtio, TxVirtio, RX_QUEUE_EVENT, RX_TAP_EVENT,
+    TX_QUEUE_EVENT,
+};
 pub use tap::{Error as TapError, Tap};
 
 #[derive(Debug)]
